// packages-ext/responses/src/generated/types.gen.ts
export interface FunctionCallOutput {
  /**
   * The unique ID of the function tool call generated by the model.
   */
  call_id: string
  /**
   * The unique ID of the function tool call output. Populated when this item is returned via API.
   */
  id: string
  output: Array<InputFileContent | InputImageContent | InputTextContent> | string
  status: FunctionCallOutputStatusEnum & unknown
  /**
   * The type of the function tool call output. Always `function_call_output`.
   */
  type: 'function_call_output'
}
// ...
export type FunctionCallOutputStatusEnum = 'completed' | 'in_progress' | 'incomplete'

export type FunctionCallStatus = 'completed' | 'in_progress' | 'incomplete'

/**
 * Function
 *
 * Defines a function in your own code the model can choose to call. Learn more about [function calling](https://platform.openai.com/docs/guides/function-calling).
 */
export interface FunctionTool {
  description: null | string
  /**
   * The name of the function to call.
   */
  name: string
  parameters: null | {
    [key: string]: unknown
  }
  strict: boolean | null
  /**
   * The type of the function tool. Always `function`.
   */
  type: 'function'
}

export interface FunctionToolChoice {
  name?: string
  type: 'function'
}

export interface FunctionToolParam {
  description?: null | string
  name: string
  parameters?: EmptyModelParam | null
  strict?: boolean
  type: 'function'
}

export type ImageDetail = 'auto' | 'high' | 'low'

export type IncludeEnum = 'message.output_text.logprobs' | 'reasoning.encrypted_content'

/**
 * Incomplete details
 *
 * Details about why the response was incomplete.
 */
export interface IncompleteDetails {
  /**
   * The reason the response could not be completed.
   */
  reason: string
}

/**
 * Input file
 *
 * A file input to the model.
 */
export interface InputFileContent {
  /**
   * The URL of the file to be sent to the model.
   */
  file_url?: string
  /**
   * The name of the file to be sent to the model.
   */
  filename?: string
  /**
   * The type of the input item. Always `input_file`.
   */
  type: 'input_file'
}

/**
 * Input file
 *
 * A file input to the model.
 */
export interface InputFileContentParam {
  file_data?: null | string
  file_url?: null | string
  filename?: null | string
  /**
   * The type of the input item. Always `input_file`.
   */
  type: 'input_file'
}

/**
 * Input image
 *
 * An image input to the model. Learn about [image inputs](/docs/guides/vision).
 */
export interface InputImageContent {
  detail: ImageDetail & unknown
  image_url: null | string
  /**
   * The type of the input item. Always `input_image`.
   */
  type: 'input_image'
}

/**
 * Input image
 *
 * An image input to the model. Learn about [image inputs](/docs/guides/vision)
 */
export interface InputImageContentParamAutoParam {
  detail?: (ImageDetail & unknown) | null
  image_url?: null | string
  /**
   * The type of the input item. Always `input_image`.
   */
  type: 'input_image'
}

/**
 * Input text
 *
 * A text input to the model.
 */
export interface InputTextContent {
  /**
   * The text input to the model.
   */
  text: string
  /**
   * The type of the input item. Always `input_text`.
   */
  type: 'input_text'
}